Refactored terrain, added wall collision
Refactored the terrain code to inherit from a terrain object base. Also implemented wall collision allowing for a new type of terrain; pillars / walls. They function as a solid block when move into from the sides, but also like a platform when fallen on from the top.

To do: Variable jump